# docs/stackr-api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Stackr Catalogue API
  version: 1.0.0
  summary: Stable public API for catalogue, delta sync and search.
servers:
  - url: https://api.stackr.app/v1
    description: Production edge gateway
  - url: http://localhost:8787/v1
    description: Local Cloudflare Worker
...
paths:
...
  /assets/manifest:
    get:
      tags: [Assets]
      operationId: getAssetManifest
      parameters:
        - name: assetType
          in: query
          schema: { type: string }
        - name: setId
          in: query
          schema: { type: string, format: uuid }
        - name: printingId
          in: query
          schema: { type: string, format: uuid }
        - name: variantId
          in: query
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Approved public catalogue asset manifest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /assets/scans/presigned-upload:
    post:
      tags: [Assets]
      operationId: createPrivateScanUpload
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mimeType, byteSize]
              properties:
                uploadId:
                  type: string
                  minLength: 8
                  maxLength: 128
                mimeType:
                  type: string
                  enum: [image/jpeg, image/png, image/webp, image/heic]
                byteSize:
                  type: integer
                  minimum: 1
                  maximum: 10485760
                expiresInSeconds:
                  type: integer
                  minimum: 60
                  maximum: 3600
      responses:
        "200":
          description: Short-lived private upload details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /assets/scans/upload:
    post:
      tags: [Assets]
      operationId: uploadPrivateScan
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
          image/heic:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Validated private scan asset record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "413":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"
...
  /admin/catalogue-ingestion/conflicts:
    get:
      tags: [Admin]
      operationId: listCatalogueIngestionConflicts
      x-stackr-internal: true
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - name: sourceId
          in: query
          schema: { type: string, format: uuid }
        - name: conflictType
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Quarantined catalogue conflicts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /admin/catalogue-ingestion/quality-report:
    get:
      tags: [Admin]
      operationId: getCatalogueQualityReport
      x-stackr-internal: true
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - name: language
          in: query
          schema: { $ref: "#/components/schemas/LanguageCode" }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Catalogue quality report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /admin/assets/migrate-existing:
    post:
      tags: [Admin]
      operationId: migrateExistingAssets
      x-stackr-internal: true
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetMigrationCommand"
      responses:
        "200":
          description: Existing asset migration accepted by the asset-processing queue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "403":
          $ref: "#/components/responses/Error"
  /admin/observability/dashboard:
    get:
      tags: [Admin]
      operationId: getProtectedObservabilityDashboard
      x-stackr-internal: true
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Latest protected health, quality, coverage, pricing, cost, provider and model snapshots.
          headers:
            Traceparent:
              $ref: "#/components/headers/Traceparent"
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservabilityDashboardResponse"
        "403":
          $ref: "#/components/responses/Error"
  /admin/observability/refresh:
    post:
      tags: [Admin]
      operationId: refreshProtectedObservabilityDashboard
      x-stackr-internal: true
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                windowHours:
                  type: integer
                  minimum: 1
                  maximum: 720
                  default: 24
      responses:
        "200":
          description: Refreshed protected dashboard snapshots.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservabilityDashboardResponse"
  /admin/observability/evaluations:
    post:
      tags: [Admin]
      operationId: storeQualityEvaluation
      x-stackr-internal: true
      security:
        - SupabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityEvaluationSubmission"
      responses:
        "201":
          description: Aggregate quality report stored without scan pixels, OCR text or user identifiers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
components:
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Supabase user access token. Service-role keys are forbidden.
  headers:
...
    Traceparent:
      schema:
        type: string
        pattern: "^00-[a-f0-9]{32}-[a-f0-9]{16}-[a-f0-9]{2}$"
    TraceId:
      schema:
        type: string
        pattern: "^[a-f0-9]{32}$"
  parameters:
...
    DeviceId:
      name: X-Stackr-Device-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
...
  responses:
    Error:
      description: Structured API error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    QualityEvaluationSubmission:
      type: object
      additionalProperties: false
      required: [runKey, manifestSha256, environment, report]
      properties:
        runKey:
          type: string
        manifestSha256:
          type: string
          pattern: "^[a-f0-9]{64}$"
        environment:
          type: string
          enum: [development, test, staging, production]
        sourceCommitSha:
          type: [string, "null"]
        report:
          type: object
          description: Aggregate stackr-quality-evaluation-v1.0.0 report. Raw scan fields are forbidden.
    ObservabilityDashboardResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [generatedAt, dashboards, releaseGates]
              properties:
                generatedAt:
                  type: string
                  format: date-time
                dashboards:
                  type: array
                  minItems: 9
                  maxItems: 9
                  items:
                    type: object
                    required: [dashboardKey, status, summary, evidenceCount, limitations]
                    properties:
                      dashboardKey:
                        type: string
                      status:
                        type: string
                        enum: [healthy, degraded, critical, unavailable]
                      summary:
                        type: object
                      evidenceCount:
                        type: integer
                        minimum: 0
                      limitations:
                        type: array
                        items: { type: string }
                latestQualityRun:
                  type: [object, "null"]
                releaseGates:
                  type: array
                  items:
                    type: object
...
    AssetMigrationCommand:
      type: object
      additionalProperties: false
      properties:
        assetType: { type: string }
        setId: { type: string, format: uuid }
        printingId: { type: string, format: uuid }
        variantId: { type: string, format: uuid }
        since: { type: string, format: date-time }
        limit: { type: integer, minimum: 1, maximum: 1000 }
        dryRun: { type: boolean }
        priority: { type: integer, minimum: 0, maximum: 100 }
        runAfter: { type: string, format: date-time }
        idempotencyKey: { type: string }
        requestId: { type: string }
...
    LanguageCode:
      type: string
      enum: [en, ja, zh-Hans, zh-Hant, ko]
...
    EnvelopeMeta:
      type: object
      required: [requestId, apiVersion, generatedAt]
      properties:
        requestId:
          type: string
        apiVersion:
          type: string
          const: "1"
        generatedAt:
          type: string
          format: date-time
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      required: [limit, nextCursor]
      properties:
        limit:
          type: integer
        nextCursor:
          type: [string, "null"]
...
    Envelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/EnvelopeMeta"
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details: {}
        meta:
          type: object